This is 3.4.5, nothing new or fixed, but contains some fixes for 64-bit targets for libmscgen.

+++ libmscgen/attribute.cpp
@@ -49,8 +49,8 @@ int CaseInsensitiveBeginsWith(const char *a, const char *b)
 
 bool CaseInsensitiveEndsWith(const char *base, const char *a)
 {
-    int i1 = strlen(a)-1;
-    int i2 = strlen(base)-1;
+    int i1 = (unsigned)strlen(a)-1;
+    int i2 = (unsigned)strlen(base)-1;
     if (i1>i2 || i1<0 || i2<0) return false;
     while(i1>=0 && i2>=0) {
         if (toupper(a[i1]) != toupper(base[i2]))
